+++ bayes_air/airport.py
@@ -2,8 +2,8 @@
 from dataclasses import dataclass
 from typing import Optional
 
-from bayes_air.components.aircraft import Aircraft
-from bayes_air.components.types import AirportCode, Time
+from bayes_air.aircraft import Aircraft
+from bayes_air.util import AirportCode, Time
 
 
 @dataclass
