+++ bayes_air/schedule.py
@@ -2,7 +2,7 @@
 from dataclasses import dataclass
 from typing import Optional
 
-from bayes_air.components.types import AirportCode, Time
+from bayes_air.util import AirportCode, Time
 
 
 @dataclass
