+++ bayes_air/util.py
